requirements.txt
streamlit
crewai
crewai-tools
python-dotenv
requests
beautifulsoup4
